requirements-2d.txt
numpy
pandas
matplotlib
seaborn
tqdm
Pillow
scikit-learn
lightgbm

# Deep learning stack (install torch/torchvision matching your CUDA)
torch
torchvision
monai

